requirements/dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    make upgrade
#
alabaster==0.7.12
amqp==1.4.9
analytics-python==1.2.9
anyjson==0.3.3
appdirs==1.4.3
astroid==2.3.3
attrs==19.3.0
babel==2.8.0
beautifulsoup4==4.8.2
billiard==3.3.0.23
bleach==3.1.0
bok-choy==1.0.1
cached-property==1.5.1
celery==3.1.26.post2
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==5.0.3
cryptography==2.8
cssselect==1.1.0
cssutils==1.0.2
ddt==1.2.2
defusedxml==0.6.0
diff-cover==2.6.0
distlib==0.3.0
git+https://github.com/django-compressor/django-appconf.git@5169ce2c92d9836e0b3ab3ec645727d9d5225d1a#egg=django-appconf
django-compressor==2.4
django-cors-headers==3.2.1
django-crispy-forms==1.8.1
django-crum==0.7.5
django-debug-toolbar==2.2
django-extra-views==0.11.0
django-filter==2.2.0
django-haystack==2.8.1
django-libsass==0.8
django-model-utils==3.2.0
django-nose==1.4.6
django-oscar-promotions==1.0.0b1
django-oscar==2.0.4
django-phonenumber-field==2.0.1
django-rest-swagger==2.2.0
django-simple-history==2.8.0
django-solo==1.1.3
django-tables2==1.21.2
django-threadlocals==0.10
django-treebeard==4.3.1
django-waffle==0.20.0
django-webtest==1.9.7
django-widget-tweaks==1.4.5
django==1.11.28
django_extensions==1.9.9
djangorestframework-csv==2.1.0
djangorestframework-datatables==0.5.1
djangorestframework-jwt==1.11.0
djangorestframework==3.6.4
docutils==0.16
drf-extensions==0.3.1
edx-auth-backends==3.0.2
edx-django-release-util==0.3.6
edx-django-sites-extensions==2.4.3
edx-django-utils==3.0
edx-drf-extensions==2.4.6
edx-ecommerce-worker==0.7.2
edx-i18n-tools==0.5.0
edx-opaque-keys==2.0.1
edx-rbac==1.1.0
edx-rest-api-client==1.9.2
edx-sphinx-theme==1.0.2
factory-boy==2.12.0
faker==4.0.1
filelock==3.0.12
freezegun==0.3.15
future==0.18.2
httpretty==0.9.7
idna==2.8
imagesize==1.2.0
importlib-metadata==1.5.0
importlib-resources==1.0.2
inflect==3.0.2
isodate==0.6.0
isort==4.3.21
itypes==1.1.0
jinja2-pluralize==0.3.0
jinja2==2.11.1
jsonfield2==3.0.3
kombu==3.0.37
lazy-object-proxy==1.4.3
lazy==1.4
libsass==0.9.2
lxml==4.5.0
markdown==2.6.9
markupsafe==1.1.1
mccabe==0.6.1
mock-django==0.6.10
mock==3.0.5
more-itertools==8.2.0
mysqlclient==1.4.6
ndg-httpsclient==0.5.1
newrelic==4.20.1.121
nose-ignore-docstring==0.2
nose==1.3.7
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.1
path.py==7.2
pathlib2==2.3.5
paypalrestsdk==1.13.1
pbr==5.4.4
phonenumbers==8.11.4
pillow==7.0.0
pluggy==0.13.1
polib==1.1.0              # via edx-i18n-tools
premailer==2.9.2
psutil==1.2.1
ptvsd==4.3.2
purl==1.5
py==1.8.1
pyasn1==0.4.8
pycodestyle==2.5.0
pycountry==17.1.8
pycparser==2.19
pycryptodomex==3.9.6
pygments==2.5.2
pyinotify==0.9.6
pyjwkest==1.3.2
pyjwt==1.7.1
pylint==2.4.4
pymongo==3.10.1
pyopenssl==19.1.0
pyparsing==2.4.6
pytest-base-url==1.4.1
pytest-html==1.22.1
pytest-metadata==1.8.0
pytest-randomly==3.2.1
pytest-selenium==1.17.0
pytest-timeout==1.3.4
pytest-variables==1.9.0
pytest==5.3.5
python-dateutil==2.8.1
python-dotenv==0.11.0
python-memcached==1.58
python-slugify==1.2.6     # via transifex-client
python3-openid==3.1.0 ; python_version >= "3"
pytz==2016.10
pyyaml==5.3
rcssmin==1.0.6
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
requests==2.22.0
responses==0.10.9
rest-condition==1.0.3
rjsmin==1.1.0
rules==2.2
sailthru-client==2.2.3
selenium==3.141.0
semantic-version==2.8.4
simplejson==3.17.0
six==1.14.0
slumber==0.7.1
snowballstemmer==2.0.0
git+https://github.com/python-social-auth/social-app-django.git@c00d23c2b45c3317bd35b15ad1b959338689cef8#egg=social-auth-app-django
social-auth-core==3.2.0
soupsieve==1.9.5
sphinx==1.5.3
sqlparse==0.3.0           # via django-debug-toolbar
stevedore==1.32.0
stripe==1.70.0
testfixtures==6.13.0
text-unidecode==1.3
toml==0.10.0
tox-battery==0.5.2
tox==3.14.5
transifex-client==0.13.7
typed-ast==1.4.1
typing==3.7.4.1
unicodecsv==0.14.1
unidecode==1.1.1          # via python-slugify
uritemplate==3.0.1
urllib3==1.25.8
virtualenv==20.0.4
waitress==1.4.3
wcwidth==0.1.8
webencodings==0.5.1
webob==1.8.6
webtest==2.0.34
wrapt==1.11.2
xss-utils==0.1.2
zeep==3.4.0
zipp==1.2.0
